Initialize members in case where SRCREV = AUTOINC
This patch fixes a case where an execption is thrown
due to referencing a uninitialized variable, moddir
in the case where SRCREV is set to AUTOINC.

In urldata_init, the ud.setup_revisons() [sic.] function is called before
some variables are initialized for the ud object.  This function
ends up requiring the moddir member when it calls srcrev_internal_helper
which tries the Hg._revision_key() method.